refactor component structure
